Add namespace for Game and Version
